lint: fix gofmt import order and unconvert in debug_test
Three golangci-lint issues from the upgrade sweep:
- cmd/thor/main.go: reorder go-isatty / uuid imports to satisfy gofmt
- tracers/js/tracer_test.go: re-align dummyStatedb method stubs via gofmt
- api/debug/debug_test.go: drop redundant uint32(^uint32(0)) conversion

+++ api/debug/debug_test.go
@@ -548,7 +548,7 @@ func initDebugServer(t *testing.T) {
 		HAYABUSA:     ^uint32(0),
 		INTERSTELLAR: ^uint32(0),
 	}
-	hayabusaTP := uint32(^uint32(0))
+	hayabusaTP := ^uint32(0)
 	thor.SetConfig(thor.Config{HayabusaTP: &hayabusaTP})
 	thorChain, err := testchain.NewWithFork(&forkConfig, 180)
 	require.NoError(t, err)
